# improve performance of 'cargo test -p codex-tui'