--- a/codex-rs/tui/src/bottom_pane/textarea.rs
+++ b/codex-rs/tui/src/bottom_pane/textarea.rs
@@ -1473,7 +1473,9 @@ mod tests {
             .timestamp() as u64;
         let mut rng = rand::rngs::StdRng::seed_from_u64(pst_today_seed);
 
-        for _case in 0..10_000 {
+        // Keep the default test fast: a few hundred cases with dozens of steps
+        // catches regressions while keeping the suite snappy.
+        for _case in 0..500 {
             let mut ta = TextArea::new();
             let mut state = TextAreaState::default();
             // Track element payloads we insert. Payloads use characters '[' and ']' which
@@ -1497,7 +1499,7 @@ mod tests {
             let mut width: u16 = rng.random_range(1..=12);
             let mut height: u16 = rng.random_range(1..=4);
 
-            for _step in 0..200 {
+            for _step in 0..60 {
                 // Mostly stable width/height, occasionally change
                 if rng.random_bool(0.1) {
                     width = rng.random_range(1..=12);
@@ -1683,6 +1685,209 @@ mod tests {
                 // After wrapping, desired height equals the number of lines we would render without scroll
                 let total_lines = total_lines as usize;
                 // state.scroll must not exceed total_lines when content fits within area height
+                if (height as usize) >= total_lines {
+                    assert_eq!(state.scroll, 0);
+                }
+            }
+        }
+    }
+
+    /// Deep fuzzing variant of `fuzz_textarea_randomized` that exercises many
+    /// more randomized operations. Ignored by default; run explicitly when
+    /// needed, e.g.:
+    ///   cargo test -p codex-tui --lib fuzz_textarea_randomized_deep -- --ignored
+    #[test]
+    #[ignore]
+    fn fuzz_textarea_randomized_deep() {
+        // Deterministic seed for reproducibility (same logic as the fast test)
+        let pst_today_seed: u64 = (chrono::Utc::now() - chrono::Duration::hours(8))
+            .date_naive()
+            .and_hms_opt(0, 0, 0)
+            .unwrap()
+            .and_utc()
+            .timestamp() as u64;
+        let mut rng = rand::rngs::StdRng::seed_from_u64(pst_today_seed);
+
+        for _case in 0..10_000 {
+            let mut ta = TextArea::new();
+            let mut state = TextAreaState::default();
+            let mut elem_texts: Vec<String> = Vec::new();
+            let mut next_elem_id: usize = 0;
+
+            let base_len = rng.random_range(0..30);
+            let mut base = String::new();
+            for _ in 0..base_len {
+                base.push_str(&rand_grapheme(&mut rng));
+            }
+            ta.set_text(&base);
+            let mut boundaries: Vec<usize> = vec![0];
+            boundaries.extend(ta.text().char_indices().map(|(i, _)| i).skip(1));
+            boundaries.push(ta.text().len());
+            let init = boundaries[rng.random_range(0..boundaries.len())];
+            ta.set_cursor(init);
+
+            let mut width: u16 = rng.random_range(1..=12);
+            let mut height: u16 = rng.random_range(1..=4);
+
+            for _step in 0..200 {
+                if rng.random_bool(0.1) {
+                    width = rng.random_range(1..=12);
+                }
+                if rng.random_bool(0.1) {
+                    height = rng.random_range(1..=4);
+                }
+
+                match rng.random_range(0..18) {
+                    0 => {
+                        let len = rng.random_range(0..6);
+                        let mut s = String::new();
+                        for _ in 0..len {
+                            s.push_str(&rand_grapheme(&mut rng));
+                        }
+                        ta.insert_str(&s);
+                    }
+                    1 => {
+                        let mut b: Vec<usize> = vec![0];
+                        b.extend(ta.text().char_indices().map(|(i, _)| i).skip(1));
+                        b.push(ta.text().len());
+                        let i1 = rng.random_range(0..b.len());
+                        let i2 = rng.random_range(0..b.len());
+                        let (start, end) = if b[i1] <= b[i2] {
+                            (b[i1], b[i2])
+                        } else {
+                            (b[i2], b[i1])
+                        };
+                        let insert_len = rng.random_range(0..=4);
+                        let mut s = String::new();
+                        for _ in 0..insert_len {
+                            s.push_str(&rand_grapheme(&mut rng));
+                        }
+                        let before = ta.text().len();
+                        let intersects_element = elem_texts.iter().any(|payload| {
+                            if let Some(pstart) = ta.text().find(payload) {
+                                let pend = pstart + payload.len();
+                                pstart < end && pend > start
+                            } else {
+                                false
+                            }
+                        });
+                        ta.replace_range(start..end, &s);
+                        if !intersects_element {
+                            let after = ta.text().len();
+                            assert_eq!(
+                                after as isize,
+                                before as isize + (s.len() as isize) - ((end - start) as isize)
+                            );
+                        }
+                    }
+                    2 => ta.delete_backward(rng.random_range(0..=3)),
+                    3 => ta.delete_forward(rng.random_range(0..=3)),
+                    4 => ta.delete_backward_word(),
+                    5 => ta.kill_to_beginning_of_line(),
+                    6 => ta.kill_to_end_of_line(),
+                    7 => ta.move_cursor_left(),
+                    8 => ta.move_cursor_right(),
+                    9 => ta.move_cursor_up(),
+                    10 => ta.move_cursor_down(),
+                    11 => ta.move_cursor_to_beginning_of_line(true),
+                    12 => ta.move_cursor_to_end_of_line(true),
+                    13 => {
+                        let payload =
+                            format!("[[EL#{}:{}]]", next_elem_id, rng.random_range(1000..9999));
+                        next_elem_id += 1;
+                        ta.insert_element(&payload);
+                        elem_texts.push(payload);
+                    }
+                    14 => {
+                        if let Some(payload) = elem_texts.choose(&mut rng).cloned()
+                            && let Some(start) = ta.text().find(&payload)
+                        {
+                            let end = start + payload.len();
+                            if end - start > 2 {
+                                let pos = rng.random_range(start + 1..end - 1);
+                                let ins = rand_grapheme(&mut rng);
+                                ta.insert_str_at(pos, &ins);
+                            }
+                        }
+                    }
+                    15 => {
+                        if let Some(payload) = elem_texts.choose(&mut rng).cloned()
+                            && let Some(start) = ta.text().find(&payload)
+                        {
+                            let end = start + payload.len();
+                            let mut s = start.saturating_sub(rng.random_range(0..=2));
+                            let mut e = (end + rng.random_range(0..=2)).min(ta.text().len());
+                            let txt = ta.text();
+                            while s > 0 && !txt.is_char_boundary(s) {
+                                s -= 1;
+                            }
+                            while e < txt.len() && !txt.is_char_boundary(e) {
+                                e += 1;
+                            }
+                            if s < e {
+                                let mut srep = String::new();
+                                for _ in 0..rng.random_range(0..=2) {
+                                    srep.push_str(&rand_grapheme(&mut rng));
+                                }
+                                ta.replace_range(s..e, &srep);
+                            }
+                        }
+                    }
+                    16 => {
+                        if let Some(payload) = elem_texts.choose(&mut rng).cloned()
+                            && let Some(start) = ta.text().find(&payload)
+                        {
+                            let end = start + payload.len();
+                            if end - start > 2 {
+                                let pos = rng.random_range(start + 1..end - 1);
+                                ta.set_cursor(pos);
+                            }
+                        }
+                    }
+                    _ => {
+                        if rng.random_bool(0.5) {
+                            let p = ta.beginning_of_previous_word();
+                            ta.set_cursor(p);
+                        } else {
+                            let p = ta.end_of_next_word();
+                            ta.set_cursor(p);
+                        }
+                    }
+                }
+
+                assert!(ta.cursor() <= ta.text().len());
+                for payload in &elem_texts {
+                    if let Some(start) = ta.text().find(payload) {
+                        let end = start + payload.len();
+                        assert_eq!(&ta.text()[start..end], payload);
+                        let c = ta.cursor();
+                        assert!(
+                            c <= start || c >= end,
+                            "cursor inside element: {}..{} at {}",
+                            start,
+                            end,
+                            c
+                        );
+                    }
+                }
+
+                let area = Rect::new(0, 0, width, height);
+                let total_lines = ta.desired_height(width);
+                let full_area = Rect::new(0, 0, width, total_lines.max(1));
+                let mut buf = Buffer::empty(full_area);
+                ratatui::widgets::WidgetRef::render_ref(&(&ta), full_area, &mut buf);
+                let _ = ta.cursor_pos(area);
+                let (_x, _y) = ta
+                    .cursor_pos_with_state(area, &state)
+                    .unwrap_or((area.x, area.y));
+                let mut sbuf = Buffer::empty(area);
+                ratatui::widgets::StatefulWidgetRef::render_ref(
+                    &(&ta),
+                    area,
+                    &mut sbuf,
+                    &mut state,
+                );
+                let total_lines = total_lines as usize;
                 if (height as usize) >= total_lines {
                     assert_eq!(state.scroll, 0);
                 }